backend/requirements.txt
django==5.2.3

djangorestframework==3.16.0

django-environ==0.12.0

psycopg2-binary==2.9.10
django-allauth
requests
celery==5.5.3   
sqlalchemy
redis==6.2.0
django-cors-headers
djangorestframework-simplejwt
dj-rest-auth
openai
google-genai
pgvector
PyGithub
astor
ruff
django-celery-beat
e2b_code_interpreter
GitPython
agno==1.7.4
psycopg[binary]